requirements.txt
pycryptodomex == 3.9.8